reduce buffer size

+++ so/so.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <math.h>
-uint32_t so_convert(uint32_t vgas, uint32_t vgas0, uint32_t vtemp, uint32_t vref) {
+uint16_t so_convert(uint32_t vgas, uint32_t vgas0, uint32_t vtemp, uint32_t vref) {
 	float vgas_f = (float)vgas*vref/4095;
 	float vgas0_f = (float)vgas0*vref/4095;
 	float vtemp_f = (float)vtemp*vref/4095;
